Commit 1/10/2019
Had to retype the pretty much the whole html document because I couldn't find out what was wrong with it, so i did my own little debug type thing. I didn't add in the script correctly and didn't add in the link for the forms css. Retyped the Customer survey because it didn't look right, beautified everything. Made the form and the field-set.