Web form exercise is ready for review
Created regular expressions for the phone number and postal code. Added a required field for the name and the email. for the css i made the validation styles where when you select the text area if its valid the color will show up green with a picture of a green check mark. If the information is invalid the text area box will show up red with a picture of an X. Pseudo class :focus.